requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antiorm==1.2.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
choicelib==0.1.5
colorama==0.4.6
coverage==7.10.2
exceptiongroup==1.3.0
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
loguru==0.7.2
msgspec==0.19.0
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.0.0
requests==2.32.4
SQLAlchemy==2.0.42
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
vbml==1.1.post1
vk_api==11.10.0
vkbottle==4.5.2
vkbottle-types==5.199.99.5
yarl==1.20.1